SQL fixes for PR #14

+++ installation/sql/postgresql/extensions.sql
@@ -896,10 +896,9 @@ CREATE INDEX "#__scheduler_tasks_idx_checked_out" ON "#__scheduler_tasks" ("chec
 -- Table structure for table "#__schemaorg"
 --
 
-CREATE TABLE IF NOT EXISTS "#__schemaorg"
-(
+CREATE TABLE IF NOT EXISTS "#__schemaorg" (
   "id" serial NOT NULL,
-  "itemId" integer,
+  "itemId" bigint,
   "context" varchar(100),
   "schemaType" varchar(100),
   "schemaForm" text,
